Feat/handle hard soft delete todo (#36)
* feat: add hard delete flag

* feat: handle hard delete flag + soft delete basic implem

* test(e2e): add e2e tests for hard/soft delete

* test(unit): add unit tests for soft/hard delete logic

+++ src/application/services/todosAppService.ts
@@ -1,5 +1,5 @@
 import { inject, singleton } from "tsyringe";
-import { Todo } from "../../domain/models/Todo";
+import { Todo, TodoState } from "../../domain/models/Todo";
 import { TodosRepository } from "../../domain/repositories/todos.repository";
 
 @singleton()
@@ -18,11 +18,18 @@ export class TodosAppService {
     return await this.todosRepository.createTodo(todo);
   }
 
+  // TODO: To refactor
   async updateTodo(todoUuid: string, ownerUuid: string, todo: Todo): Promise<Todo> {
     return await this.todosRepository.updateTodo(todo);
   }
 
-  async deleteTodo(todoUuid: string, ownerUuid: string): Promise<void> {
-    return await this.todosRepository.deleteTodo(todoUuid, ownerUuid);
+  async deleteTodo(todoUuid: string, ownerUuid: string, hardDelete = false): Promise<void> {
+    if (hardDelete) {
+      return await this.todosRepository.deleteTodo(todoUuid, ownerUuid);
+    }
+    // TODO: Would be likely better in a transaction to avoid concurrency issues
+    const todo = await this.getTodo(todoUuid, ownerUuid);
+    todo.state = TodoState.DELETED;
+    await this.updateTodo(todoUuid, ownerUuid, todo);
   }
 }
